Use halfs in G-Buffer

+++ Engine/Shaders/Common/Material.h
@@ -7,8 +7,8 @@
 #include <metal_stdlib>
 
 struct Material {
-    metal::texture2d<float> albedo [[id(0)]];
-    metal::texture2d<float> roughness [[id(1)]];
+    metal::texture2d<half> albedo [[id(0)]];
+    metal::texture2d<half> roughness [[id(1)]];
     metal::texture2d<float> normals [[id(2)]];
-    metal::texture2d<float> metallic [[id(3)]];
+    metal::texture2d<half> metallic [[id(3)]];
 };
